+++ src/mobile/lib/app/utils/widgets/card_produto_search.dart
@@ -66,16 +66,19 @@ class CardProdutoSearch extends StatelessWidget {
                         overflow: TextOverflow.ellipsis,
                       ),
                       RichText(
-                          text: TextSpan(text: "R\$ ", style: Theme.of(context).primaryTextTheme.displayMedium, children: [
-                        TextSpan(
-                          text: UtilBrasilFields.obterReal(item.preco, moeda: false),
-                          style: Theme.of(context).primaryTextTheme.bodyLarge,
-                        ),
-                        TextSpan(
-                          text: ' / ${item.unidadeMedida}',
-                          style: Theme.of(context).primaryTextTheme.displayMedium,
-                        )
-                      ])),
+                          text: TextSpan(
+                              text: "R\$ ",
+                              style: Theme.of(context).primaryTextTheme.displayMedium,
+                              children: [
+                            TextSpan(
+                              text: UtilBrasilFields.obterReal(item.preco.toDouble(), moeda: false),
+                              style: Theme.of(context).primaryTextTheme.bodyLarge,
+                            ),
+                            TextSpan(
+                              text: ' / ${item.unidadeMedida}',
+                              style: Theme.of(context).primaryTextTheme.displayMedium,
+                            )
+                          ])),
                       Padding(
                         padding: const EdgeInsets.only(top: 4.0),
                         child: Row(
@@ -90,7 +93,8 @@ class CardProdutoSearch extends StatelessWidget {
                             ),
                             RichText(
                               text: TextSpan(
-                                text: "${UtilBrasilFields.obterReal(item.rating.toDouble(), moeda: false, decimal: 1)} ",
+                                text:
+                                    "${UtilBrasilFields.obterReal(item.rating.toDouble(), moeda: false, decimal: 1)} ",
                                 style: Theme.of(context).primaryTextTheme.bodyLarge,
                                 children: [
                                   TextSpan(
@@ -157,7 +161,9 @@ class CardProdutoSearch extends StatelessWidget {
                           favoritoController.isFavorito = !favoritoController.isFavorito;
                         }
                       },
-                      icon: favoritoController.isFavorito ? Image.asset('assets/fav_red.png') : Image.asset('assets/fav_grey.png'),
+                      icon: favoritoController.isFavorito
+                          ? Image.asset('assets/fav_red.png')
+                          : Image.asset('assets/fav_grey.png'),
                     );
                   })
                 ],
